Attempt to fix warning
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@8196 9d14d920-75d5-0310-8bd7-a52572981aee

+++ engine/tester/iters.c
@@ -145,7 +145,7 @@ test_param_iteration_clone(const test_param_iteration *i)
     test_param           *p;
     test_param           *pc;
 
-    ENTRY("0x%x", (unsigned int)i);
+    ENTRY("0x%x", (size_t)i);
 
     if (ic == NULL)
     {
